api/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
mysql-connector==2.2.9
requests==2.22.0
python-dotenv==0.10.1